The package generated by log_collector doesn't contain the Redis Enterprise Support Package
This is because we use the full path of the generated SP (including the leading "/tmp") when we build the output path. Instead, for building the output path, we should only use the SP file name. There are various ways to fix this. I went with a second capture group.